+++ litestar/dto/factory/abc.py
@@ -1,22 +1,19 @@
 from __future__ import annotations
 
 from abc import ABCMeta, abstractmethod
-from collections.abc import Collection as CollectionsCollection
 from itertools import chain
 from typing import TYPE_CHECKING, Generic, TypeVar
 
-from msgspec import Struct
-
+from litestar.dto.factory.backends import MsgspecDTOBackend
 from litestar.dto.interface import DTOInterface
 from litestar.types.builtin_types import NoneType
 from litestar.utils.signature import ParsedType
 
-from .backends import MsgspecDTOBackend
 from .config import DTOConfig
 from .exc import InvalidAnnotation
 from .field import Mark, Purpose
 from .types import FieldDefinition, FieldDefinitionsType, NestedFieldDefinition
-from .utils import get_model_type_hints, parse_configs_from_annotation
+from .utils import parse_configs_from_annotation
 
 if TYPE_CHECKING:
     from typing import Any, ClassVar, Collection, Generator, TypeAlias
@@ -25,17 +22,14 @@ if TYPE_CHECKING:
 
     from litestar.connection import Request
     from litestar.handlers import BaseRouteHandler
+    from litestar.types.serialization import LitestarEncodableType
 
     from .backends import AbstractDTOBackend
 
-__all__ = [
-    "AbstractDTOFactory",
-    "MsgspecBackedDTOFactory",
-]
+__all__ = ["AbstractDTOFactory"]
 
 AnyRequest: TypeAlias = "Request[Any, Any, Any]"
 DataT = TypeVar("DataT")
-StructT = TypeVar("StructT", bound=Struct)
 
 
 class AbstractDTOFactory(DTOInterface, Generic[DataT], metaclass=ABCMeta):
@@ -50,8 +44,6 @@ class AbstractDTOFactory(DTOInterface, Generic[DataT], metaclass=ABCMeta):
     """Config objects to define properties of the DTO."""
     model_type: ClassVar[type[Any]]
     """If ``annotation`` is an iterable, this is the inner type, otherwise will be the same as ``annotation``."""
-    dto_backend_type: ClassVar[type[AbstractDTOBackend]]
-    """DTO backend type."""
 
     _reverse_field_mappings: ClassVar[dict[str, FieldDefinition]]
     _type_config_backend_map: ClassVar[dict[tuple[ParsedType, DTOConfig], AbstractDTOBackend]]
@@ -100,6 +92,10 @@ class AbstractDTOFactory(DTOInterface, Generic[DataT], metaclass=ABCMeta):
         """Return the data held by the DTO."""
         return self._data
 
+    def to_encodable_type(self) -> LitestarEncodableType:
+        backend = self.get_backend(Purpose.READ, self._connection.route_handler)
+        return backend.encode_data(self._data, self._connection)
+
     @classmethod
     def from_data(cls, data: DataT | Collection[DataT], connection: AnyRequest) -> Self:
         """Construct an instance from data.
@@ -112,6 +108,22 @@ class AbstractDTOFactory(DTOInterface, Generic[DataT], metaclass=ABCMeta):
             AbstractDTOInterface instance.
         """
         return cls(data=data, connection=connection)
+
+    @classmethod
+    def from_bytes(cls, raw: bytes, connection: AnyRequest) -> Self:
+        """Construct an instance from bytes.
+
+        Args:
+            raw: Raw connection data.
+            connection: A byte representation of the DTO model.
+
+        Returns:
+            AbstractDTOFactory instance.
+        """
+        backend = cls.get_backend(Purpose.WRITE, connection.route_handler)
+        return cls(
+            data=backend.populate_data_from_raw(cls.model_type, raw, connection.content_type[0]), connection=connection
+        )
 
     @classmethod
     @abstractmethod
@@ -275,7 +287,7 @@ class AbstractDTOFactory(DTOInterface, Generic[DataT], metaclass=ABCMeta):
             if backend is None:
                 backend = cls._type_config_backend_map.setdefault(
                     key,
-                    cls.dto_backend_type.from_field_definitions(parsed_type, cls.parse_model(cls.model_type, config)),
+                    MsgspecDTOBackend.from_field_definitions(parsed_type, cls.parse_model(cls.model_type, config)),
                 )
 
             if is_data_type:
@@ -297,149 +309,3 @@ class AbstractDTOFactory(DTOInterface, Generic[DataT], metaclass=ABCMeta):
         """
         config = cls.configs[0]
         return cls._handler_config_backend_map[(purpose, route_handler, config)]
-
-
-class MsgspecBackedDTOFactory(AbstractDTOFactory[DataT], Generic[DataT], metaclass=ABCMeta):
-    dto_backend_type = MsgspecDTOBackend
-
-    @classmethod
-    def from_bytes(cls, raw: bytes, connection: AnyRequest) -> Self:
-        """Construct an instance from bytes.
-
-        Args:
-            raw: Raw connection data.
-            connection: A byte representation of the DTO model.
-
-        Returns:
-            AbstractDTOFactory instance.
-        """
-        backend = cls.get_backend(Purpose.WRITE, connection.route_handler)
-        parsed = backend.parse_raw(raw, connection.content_type[0])
-        return cls(
-            data=_build_data_from_struct(cls.model_type, parsed, backend.field_definitions), connection=connection
-        )
-
-    def to_encodable_type(self) -> Any:
-        backend = self.get_backend(Purpose.READ, self._connection.route_handler)
-        if isinstance(self._data, CollectionsCollection):
-            return backend.parsed_type.origin(
-                _build_struct_from_model(datum, backend.data_container_type) for datum in self._data  # pyright:ignore
-            )
-        return _build_struct_from_model(self._data, backend.data_container_type)
-
-
-def _build_model_from_struct(model_type: type[DataT], data: Struct, field_definitions: FieldDefinitionsType) -> DataT:
-    """Create instance of ``model_type``.
-
-    Args:
-        model_type: the model type received by the DTO on type narrowing.
-        data: primitive data that has been parsed and validated via the backend.
-        field_definitions: model field definitions.
-
-    Returns:
-        Data parsed into ``model_type``.
-    """
-    unstructured_data = {}
-    for k in data.__slots__:  # type:ignore[attr-defined]
-        v = getattr(data, k)
-
-        field = field_definitions[k]
-
-        if isinstance(field, NestedFieldDefinition) and isinstance(v, CollectionsCollection):
-            parsed_type = field.field_definition.parsed_type
-            if parsed_type.origin is None:  # pragma: no cover
-                raise RuntimeError("Unexpected origin value for collection type.")
-            unstructured_data[k] = parsed_type.origin(
-                _build_model_from_struct(field.nested_type, item, field.nested_field_definitions) for item in v
-            )
-        elif isinstance(field, NestedFieldDefinition) and isinstance(v, Struct):
-            unstructured_data[k] = _build_model_from_struct(field.nested_type, v, field.nested_field_definitions)
-        else:
-            unstructured_data[k] = v
-
-    return model_type(**unstructured_data)
-
-
-def _build_data_from_struct(
-    model_type: type[DataT], data: Struct | Collection[Struct], field_definitions: FieldDefinitionsType
-) -> DataT | Collection[DataT]:
-    """Create instance or iterable of instances of ``model_type``.
-
-    Args:
-        model_type: the model type received by the DTO on type narrowing.
-        data: primitive data that has been parsed and validated via the backend.
-        field_definitions: model field definitions.
-
-    Returns:
-        Data parsed into ``model_type``.
-    """
-    if isinstance(data, CollectionsCollection):
-        return type(data)(  # type:ignore[return-value]
-            _build_data_from_struct(model_type, item, field_definitions) for item in data  # type:ignore[call-arg]
-        )
-    return _build_model_from_struct(model_type, data, field_definitions)
-
-
-def _build_struct_from_model(model: Any, struct_type: type[StructT]) -> StructT:
-    """Convert ``model`` to instance of ``struct_type``
-
-    It is expected that attributes of ``struct_type`` are a subset of the attributes of ``model``.
-
-    Args:
-        model: a model instance
-        struct_type: a subclass of ``msgspec.Struct``
-
-    Returns:
-        Instance of ``struct_type``.
-    """
-    data = {}
-    for key, parsed_type in get_model_type_hints(struct_type).items():
-        model_val = getattr(model, key)
-        if parsed_type.is_subclass_of(Struct):
-            data[key] = _build_struct_from_model(model_val, parsed_type.annotation)
-        elif parsed_type.is_union:
-            data[key] = _handle_union_type(parsed_type, model_val)
-        elif parsed_type.is_collection:
-            data[key] = _handle_collection_type(parsed_type, model_val)
-        else:
-            data[key] = model_val
-    return struct_type(**data)
-
-
-def _handle_union_type(parsed_type: ParsedType, model_val: Any) -> Any:
-    """Handle union type.
-
-    Args:
-        parsed_type: Parsed type.
-        model_val: Model value.
-
-    Returns:
-        Model value.
-    """
-    for inner_type in parsed_type.inner_types:
-        if inner_type.is_subclass_of(Struct):
-            # If there are multiple struct inner types, we use the first one that creates without exception.
-            # This is suboptimal, and perhaps can be improved by assigning the model that the inner struct
-            # was derived upon to the struct itself, which would allow us to isolate the correct struct to use
-            # for the nested model type instance. For the most likely case of an optional union of a single
-            # nested type, this should be sufficient.
-            try:
-                return _build_struct_from_model(model_val, inner_type.annotation)
-            except (AttributeError, TypeError):
-                continue
-    return model_val
-
-
-def _handle_collection_type(parsed_type: ParsedType, model_val: Any) -> Any:
-    """Handle collection type.
-
-    Args:
-        parsed_type: Parsed type.
-        model_val: Model value.
-
-    Returns:
-        Model value.
-    """
-    if parsed_type.inner_types and (inner_type := parsed_type.inner_types[0]).is_subclass_of(Struct):
-        return parsed_type.origin(_build_struct_from_model(m, inner_type.annotation) for m in model_val)
-    return model_val
