+++ litestar/dto/factory/backends/abc.py
@@ -9,16 +9,19 @@ from typing import TYPE_CHECKING, Generic, TypeVar
 from .utils import build_annotation_for_backend
 
 if TYPE_CHECKING:
-    from typing import Any
+    from typing import Any, Collection
 
     from typing_extensions import Self
 
+    from litestar.connection import Request
     from litestar.dto.factory.types import FieldDefinitionsType
     from litestar.enums import MediaType
+    from litestar.types.serialization import LitestarEncodableType
     from litestar.utils.signature import ParsedType
 
 __all__ = ("AbstractDTOBackend",)
 
+T = TypeVar("T")
 BackendT = TypeVar("BackendT")
 
 
@@ -70,4 +73,26 @@ class AbstractDTOBackend(ABC, Generic[BackendT]):
 
         Returns:
             The raw bytes parsed into primitive python types.
+        """
+
+    @abstractmethod
+    def populate_data_from_raw(self, model_type: type[T], raw: bytes, media_type: MediaType | str) -> T | Collection[T]:
+        """Parse raw bytes into instance of `model_type`.
+
+        Args:
+            model_type: Type of model to populate.
+            raw: bytes
+            media_type: encoding of raw data
+        """
+
+    @abstractmethod
+    def encode_data(self, data: Any, connection: Request[Any, Any, Any]) -> LitestarEncodableType:
+        """Encode data into a ``LitestarEncodableType``.
+
+        Args:
+            data: Data to encode.
+            connection: Connection - can be used for content negotiation.
+
+        Returns:
+            Encoded data.
         """
